refactoring, created separate files
